refac(linkAddEdit): refactoring
